Consolidate dep filtering in appropriate step
This was an existing configuration feature that I forgot to bring across
to the new dedicated dependency step.
