+++ .changeset/fair-teachers-sleep.md
@@ -1,0 +1,5 @@
+---
+'skuba': patch
+---
+
+**configure:** List filtered devDependencies upfront
