+++ src/cli/configure/dependencies/index.ts
@@ -1,4 +1,5 @@
 export { seekDatadogCustomMetrics } from './seekDatadogCustomMetrics';
 export { seekKoala } from './seekKoala';
 export { skuba } from './skuba';
+export { skubaDeps } from './skubaDeps';
 export { skubaDive } from './skubaDive';
