+++ src/cli/configure/dependencies/skubaDeps.test.ts
@@ -1,0 +1,28 @@
+import { skubaDeps } from './skubaDeps';
+
+describe('skubaDeps', () => {
+  it('strips bundled dev dependencies', () => {
+    const input = {
+      dependencies: {
+        'skuba-dive': '1.0.0',
+      },
+      devDependencies: {
+        eslint: '0.0.1',
+        'pino-pretty': '0.0.1',
+        typescript: '0.0.1',
+      },
+    };
+
+    const result = skubaDeps(input);
+
+    expect(result).toHaveLength(0);
+    expect(input).toEqual({
+      dependencies: {
+        'skuba-dive': '1.0.0',
+      },
+      devDependencies: {
+        'pino-pretty': '0.0.1',
+      },
+    });
+  });
+});
